fix: overwrite xdg state for installs on fetching spec (#3339)
Summary:
Some steps of install add to the install spec. So we need to be careful
about overwriting with less information.

This PR simplifies things. Whenever we fetch an install spec, we erase
everything else and start over. That means that start will add things
and then write. Nothing should change the files after that.

However bix bootstrap takes in a file path. So this adds a command to
get the slug for an install spec. Then we use that everywhere.

Test Plan:
bix bootstrap.

CI

+++ bi/pkg/installs/write.go
@@ -37,11 +37,16 @@ func (env *InstallEnv) Remove() error {
 
 func (env *InstallEnv) WriteSpec(force bool) error {
 	specPath := env.SpecPath()
-
-	if force {
-		// Remove the old file if it exists
-		_ = os.Remove(specPath)
+	if _, err := os.Stat(specPath); err == nil {
+		if force {
+			slog.Debug("Removing old spec", slog.String("path", specPath))
+			_ = os.Remove(specPath)
+		} else {
+			slog.Debug("Spec already exists", slog.String("path", specPath))
+			return nil
+		}
 	}
+
 	// Write the spec
 	err := env.Spec.WriteToPath(specPath)
 	if err != nil {
@@ -53,9 +58,17 @@ func (env *InstallEnv) WriteSpec(force bool) error {
 func (env *InstallEnv) WriteSummary(force bool) error {
 	summaryPath := env.SummaryPath()
 
-	if force {
-		// Remove the old file if it exists
-		_ = os.Remove(summaryPath)
+	if _, err := os.Stat(summaryPath); err == nil {
+		// If the file exists then we don't need to write it
+		if force {
+			slog.Debug("Removing old summary", slog.String("path", summaryPath))
+			// Remove the old file if it exists
+			_ = os.Remove(summaryPath)
+		} else {
+			slog.Debug("Summary already exists", slog.String("path", summaryPath))
+			return nil
+
+		}
 	}
 	err := env.Spec.WriteStateSummary(summaryPath)
 	if err != nil {
@@ -67,9 +80,14 @@ func (env *InstallEnv) WriteSummary(force bool) error {
 func (env *InstallEnv) WriteKubeConfig(force bool) error {
 	kubeConfigPath := env.KubeConfigPath()
 
-	if force {
-		// Remove the old file if it exists
-		_ = os.Remove(kubeConfigPath)
+	if _, err := os.Stat(kubeConfigPath); err == nil {
+		if force {
+			slog.Debug("Removing old kubeconfig", slog.String("path", kubeConfigPath))
+			_ = os.Remove(kubeConfigPath)
+		} else {
+			slog.Debug("Kubeconfig already exists", slog.String("path", kubeConfigPath))
+			return nil
+		}
 	}
 
 	// TODO
